.dockerignore: switch to allow list rather than deny list
Change our .dockerignore to use an allow list pattern where everything
is denied by default and then `!` rules are used to allow individual
files and directories to be copied to the Docker demon on image
creation. Ideally this will help keep us from accidentally leaking
secrets or random junk into the images, especially when building in
a development environment.