dev: Fix `make init` by removing wait dependency between web and nodejs
The `-wait tcp://nodejs:8001` dependency for the web container on the
nodejs container added in 09beab5 made it easier to tell when the stack
is actually ready to handle http requests. Unfortunately, it also broke
the `make init` command by making the web container unstable for
a variable amount of time (up to 5 minutes in my local tests) depending
on the state of the node_modules and the complexity of the initial
webpack compile.

Change-Id: Icbe66357f97e22feb330cbdd0d9f4944a7b6e1c8