ui: Use navigation drawer rather than tabs
Replace the tabbed site navigation embedded in the app bar with
a navigation drawer. The navigation links in the drawer are dynamically
created based on the routes registered with the router.

The navigation drawer is set to display in a collapsed state showing
just an icon for each item by default on viewports that are medium width
and wider. In this state it will expand on hover to show labels for the
icons. On small and xsmall viewports, the drawer is hidden by default
and shown at its full width when the "hamburger" nav icon is clicked.

A placeholder footer is also added in this patch. Eventually I assume
that we will use it for adding links to things like a privacy policy.

Change-Id: I1c487d89b81f2a5760f89d917d73303b031d19d6