ui: redesign user status component
Redesign our UserStatus component to use rounded buttons in both
authenticated and anonymous state and to show a menu of options when
authenticated. Currently we don't have anything to add to the
authenticated user's menu other than a logout action, but that is
expected to change as we add additional screens and functionality.

Change-Id: I4b259c0ac0d8d5d0e62a5e1c61f15f0585628532