toolinfo schemas: Fix 1.1.1 and 1.2.0 definitions
The 1.1.1 and 1.2.0 schemas are updated to place definitions for reused
types such as "url", "wiki", "language" at the proper level in the
schema. The prior placement was incorrect due to an off-by-one indenting
error in the canonical YAML configuration file. This was noticed while
attempting to use the 1.2.0 schema to generate Django database models.

Change-Id: I9f9cac7dbff26cfdc582bb262f0949cc6e413220