cleanup: restructure Django apps packages
Django's system of "projects" and "applications" has been a source of
confusion for many over the years. In an attempt to keep our source tree
a bit cleaner as the project grows we are adopting the pattern of making
an `apps` package inside the `toolhub` project and placing our
"applications" inside that package.

The `vue` application is left at the root of the repo and outside the
`toolhub` project entirely. This currently seems like a reasonable
compromise to keep the Vue frontend more visible in the repo. It also
saves us from a lot of changes to tooling and configuration that moving
the Vue code would require.

Change-Id: I966240052e4a9dd4014a93ff90263de50a0ec1e0