ci: Trick pipelinelib into actually running tests
While comparing local and CI output I noticed that several of the
containers were being built on the CI side but apparently not run. The
stages containing runtime environment variable injection were running,
but the others were not. This change adds a "PIPELINE_RUNNING=1" env var
to every stage. This does seem to be needed magic to trigger the
container execution.

Change-Id: I4c08d98a14ea1547fc15cd7c1cfa1c1b7baea41f