api: Move DRF code into apps
It has become apparent that keeping all of the Django Rest Framework
views and serializers in a central 'api' app is not going to scale well.
This patch moves the views and serializers to the same apps which hold
the models that they operate on. The remaining bits of the api app are
moved into the core project.

Change-Id: I4c8d9df908efbd157ed92a3de321255901d6fb54