Add a basic crawler
The new toolhub.apps.crawler.tasks.Crawler class is a start on a web
crawler for collecting toolinfo data from external URLs. The initial
implementation is missing quite a few "nice to have" features including
rate limiting and robust input validation. It does however deal
gracefully with crawling many existing published toolinfo URLs from
the configuration for Hay's Directory.

A `./manage.py crawl` command is included to trigger a run of the
crawler. This can be used from the development environment like this:

  $ make web-shell
  $ poetry run ./manage.py crawl

Bug: T264810
Change-Id: Ic1ef8fb991a9661dfc60dfd9c39796c665e87fb1