api/tools: Expand icon to include link to media
Add a custom serializer for a tool's 'icon' attribute. The input and
stored value is a Commons File namespace url. This is reasonable for
input, but api consumers should not have to know esoteric things about
MediaWiki internals to turn that content page URL into an href that will
fetch the image described in the File page.

This implementation "cheats" by using Special:FilePath and the File's
title. The MediaWiki special page will redirect to the original media.
If we find that we really want more information on the linked media
(dimensions, license, etc), a reasonable future addition would be to
call the MediaWiki action=query&prop=imageinfo API at the time the data
is stored into the database and persist that in the db.

Bug: T268789
Change-Id: I544f7bfb30bf6977502abdc2753b3c2cb32c3a46