crawler: Normalize bare values as arrays for storage
Several fields in the toolinfo.json schema are defined as "oneOf"
selections which allow either single value or an array of values. This
flexibility is nice for making it as easy as possible to create a valid
toolinfo record, but preserving this flexibility in the storage and
output layers puts an extra burden on consumers of the crawled records.
We can remove some of this burden by normalizing these fields to always
be arrays in the storage and output layers. This should make it slightly
easier for consumers to reason about the structure of a toolinfo record.

Bug: T268789
Change-Id: Id651d329cea468d8b00d9f82714ea87d913bf90b