crawler: Fix logic bug rejecting too many toolinfo records
A logic bug was lurking in the original crawler added in d728082 which
rejected all subsequent records in an array input following the first
invalid record. This occurred because the valid status of the entire
crawled URL was being used to decide if we should skip storing the
current record or not rather than using a local check for the exact
record being processed.

This logic had been corrected to use a local validity flag that is reset
to True on each iteration of the toolinfo processing loop. This status
flag is then used to determine if the validity of the URL should be
changed (any invalid records is recorded as an invalid URL) and also if
processing should skip to the next tool in the list of toolinfo records
collected from the URL.

Change-Id: Iceabf50862679b80bdaf714a2b407c9cc5da78d2