rtl: Use RTL aware css helper classes
Vuetify comes with a number of handy css classes for adjusting
horizontal and vertical spacing. These are documented at
<https://v2.vuetifyjs.com/en/styles/spacing/> and generally follow the
naming convention of "{property}{direction}-{size}". The "direction"
designates which side of the affected element is adjusted. When creating
layouts that are easily mirrored for RTL languages, rather than using
"l" and "r" for "left" and "right", we should use "s" and "e" for
"start" and "end". The "l" and "r" directions are physical, but the "s"
and "e" directions are logical and mirror as desired when $vuetify.rtl
is toggled.

It would be nice to add a linter check that warns us when we
accidentally use an "l" or "r" class in the future.

Bug: T269056
Change-Id: Ifcc1ec547ad853f12bb0c75aacbf093e68f819b8