auditlog: Ignore update events for user models
The configuration we are using for OAuth authentication will touch the
ToolhubUser model each time a user logs in. These events are not of any real
value in the auditlog feed, so we are adding a guard condition in the
model update handler to ignore update events for user models. If we add
more fine grained actions in the future, it might be reasonable to
change this guard and emit a log if a user is renamed because of
a change in their SUL account.

Bug: T264809
Change-Id: Ia9e6eebfa4655b926f11154cc1de3cb0056a0d4e