pipeline: Build demo server using Blubber
Add new Blubber config to generate an image suitable for use as a stand
alone demo server for Toolhub. This image is built up using several new
Blubber variants:

; prep-nodejs
: Compiles Vue frontend code to static js and css in vue/dist.
; prep
: Copies in vue/dist from prep-nodejs and then runs Django's
  collectstatic and compilemessages management commands to produce
  production ready static asset and localization files.
; demo
: Copies in all of /srv/app from prep and then runs Django with a local
  sqlite3 database and initial seed data.

A few settings defaults have also been changed to simplify the number of
environment overrides that are needed for running tests and the demo
server.

Bug: T269018
Change-Id: I1ef629001d3cf701ebb30faad85a3b46ecb7a586