crawler: Move create/update logic to custom manager
Move the business logic for cleaning and then creating/updating a Tool
model from toolinfo data to a new ToolManager.from_toolinfo()
function. This puts the logic closer to the Tool model definition and
makes it easier to reuse if needed in the future.

The new method adds a few features over the prior implementation:
* Tool names are now 'slugified' using Django's standard logic.
* Tool names generated by Striker are given special treatment.
* Existing records are only updated if their user managed data has changed.

Bug: T264810
Change-Id: I4e37e9f1175ca4a7a61fdc0ef2a4fc5042babe02