crawler: track total and updated tools
Count the number of updated and total tools while crawling. These new
counters are exposed in the API. The total tool count has been added to
the crawler history view. The graph of new tools on that page has been
replaced with a graph of total tools. The number of new and updated
tools from the most recent run is used on the Home view.

Change-Id: Ib7db902d5672acf132523420a21b25c8ce029c42