i18n: Configure fallback locales for vue-i18n
Sometimes there will be gaps in our translation data for a given locale.
Vue-i18n allows us to configure a list of "fallback" locales to lookup
messages that are not found so that we can provide some message back to
the reader. The most trivial form of this is setting a single locale to
be the fallback for all others, which is what we have been doing up to
now.

MediaWiki has a well established and maintained set of rules for doing
this same sort of lookup cascade. Thanks to some notes left behind by
@Nikerabbit at <https://phabricator.wikimedia.org/T258492#6490574> we
can extract this configuration from MediaWiki and use it to configure
vue-i18n.

Bug: T270671
Change-Id: Ieb4857abac04fe9eadee1ec04be839149bd8c988