user.Status: Remove invalid property references from login button
In e3a2f42 I added two data properties to the user.Status component to
stop warnings about undefined variables I was seeing in my browser's
console. Later I noticed that eslint was complaining that these new
variables were shadowing variables set at a higher scope. On deeper
inspection I realized that the login button contained copy-n-paste
attributes for v-bind and v-on that do not belong there. These bindings
are used in the other button attached to the component to toggle a menu
state using the menu's scoped slot named "activator". They are not
needed on the login button which does not control a similar menu
display.

Change-Id: Ie51031948a32bfbc5488481af283cc95adf8230f