i18n: Create a qqx locale
MediaWiki has an amazingly useful feature for figuring our which
messages are used in a given page: `?uselang=qqx`. This fake language
code renders messages as `(message-name)` in the UI. This in turn makes
it much easier to see where a given message is placed in the UI.

We can replicate this functionality with our vue-i18n library in the
frontend by programmatically generating translations for the qqx locale
based on the keys found in our en locale. The keys are a direct copy and
the values are the key names surrounded by parens to match the MediaWiki
output.

Bug: T269164
Change-Id: I8fb20ee07cc038dce7aaa1a0e4f671277cb18e03