home: use flexbox to layout tool cards
Getting breakpoints, widths, and paddings correct to make a grid display
of many equal sized cards is tricky with classic css. Luckily the folks
who make browser rendering engines and standards for them have
recognized this and added flexbox layout. With flexbox we can delegate
a lot of tricky work to the rendering engine itself.

* Setup column breakpoints that will show 1 card per row on tiny
  viewports, 2 per row on small viewports, 3 per row on medium, and
  4 per row on large viewports.
* Add css rules for a "home-tool-card" which uses flexbox to control the
  vertical spacing of sections within the card.
* Break with material design convention by stacking card action buttons
  vertically so that we can keep both the icon and longer text even at
  the tiny breakpoint.

There are still some visual quirks in this layout, but the regularity of
the grid is much improved--especially at the medium and smaller
breakpoints.

Co-authored-by: Srishakatux <ssethi@wikimedia.org>
Change-Id: I5945726ea1cac6e7e05fd89a3246ec989a1003ef