i18n: Consolidate locale change logic in a Vuex module
Consolidate logic and code to setup $i18n.locale and $vuetify.rtl in the
locale Vuex module. In the prior setup we had code in both App.vue and
SelectLocale.vue and no easy way to call either chunk of logic from
elsewhere in the application if desired. The Vuex module allows us to
both consolidate logic so that it is easier to debug as well as exposing
relatively simple to use Vuex actions that can be called from any Vue
component, including other Vuex actions.

This Vuex module is using a slightly different coding style from
existing modules. The module exports it's getters, actions, and
mutations independent of the default export to make unit testing more
straight forward. In a related first, the module has tests!

Change-Id: Ib432f47d0cffccbddcc5bfed70f0294d1c4d7372