i18n: Make user preferred content languages easier to find
In 9da4a4c the locale select list was updated to sort English to the
top. The idea there was to make it easier for users to find the language
in the long list of options. The reasoning here was great, but the
implementation feels heavy handed. Instead of always sorting English to
the top, we can ask the browser which content languages the user has
configured their browser to request and sort these to the top instead.

Change-Id: I7c0e80c7707a1ab9a8387b208e6348e2455c199f