toolinfo: Track toolinfo origin (crawler vs api)
Add a new "origin" field to our Tool model and set its value based on
the input method used to submit the edit. Once an origin has been set it
is considered a validation error to change it. This will be used to
ensure that tools maintained by the crawler and API input are separated
in the future to prevent editing of a crawler created record via the
API.

Bug: T271366
Change-Id: If64073b5f86dc4b03bb9795ea531a11f6c804987