ui: Filter url_multilingual links by active locale
Some toolinfo url values (for example 'developer_docs_url') may contain
multiple urls differing by language. Rather than displaying all urls, we
filter these arrays of multilingual urls to select the url which is the
best match for the current locale. The vue-i18n plugin is queried for a list
of fallback locales and the first url with a matching language is
selected.

Change-Id: Ia828a02194ffd1f668292aea2a49b393606668f9