api: Add SPDX endpoints
Add new API endpoints:
* GET /api/spdx/ - List all known SPDX licenses
* GET /api/spdx/{id}/ - Info for a specific SPDX license

The backing data for these API endpoints comes from the
spdx-license-list python library which in turn aggregates data from
<https://github.com/spdx/license-list-data>.

In addition to the API endponts, we are also adding a custom validator
and attaching it to the Tool.license field. This is deliberately not
using Django's built-in `choices=...` functionality to avoid polluting
our model migrations and OpenApi schemas with the ~450 license choices.

Bug: T273670
Change-Id: I2888b4c548c41d69c44c76dd72c87a90ee5e5946