toolinfo: Reduce serializer config duplication
Use Meta class inheritance and the somewhat obscure technique of class
scope initialization code to share the field list of
CreateToolSerializer.Meta with UpdateToolSerializer.Meta. The fields
list needed in UpdateToolSerializer is a strict subset of the fields
used in CreateToolSerializer which allows us to copy the parent list at
runtime and remove the unwanted elements. This is hoped to be easier to
understand at the code level than visually diffing the prior explicit
lists.

The meta programming technique used here is to initialize member
variables in the child class by copying and then modifying related
values from the parent class. In this particular case that copy + modify
can be done in a single statement. The general technique however can be
extended in more complex usage by inserting additional statements in the
class scope. These statements will be evaluated by the Python runtime in
a special namespace during the normal course of constructing the class
object. See <https://stackoverflow.com/a/51324825/8171> for a slightly
better explanation.

Change-Id: Ib2a71dd3b42d09870abf255604e0722296dd2372