pipeline: Hack around cryptography rust build requirement
Temporarily work around the requirement of a rust toolchain for building
the python cryptography >=3.4 library by pinning to an older version of
the docker-registry.wikimedia.org/python3-build-buster Docker base image
which includes a version of pip new enough to automatically install from
the upstream provided binary wheel.

This is a HACK! This should not be the end state of addressing this bug,
but it should allow us to continue running CI jobs until a better long
term solution can be implemented.

Bug: T274310
Change-Id: I8b9cce9ab7d2eedec447892f2b9c95dbf6bb76bb