auditlog: Replace Tool id with name in api response
Our Tool model has a numeric primary key, but it also has a unique
"name" which we use in API calls. When serializing an auditlog entry
related to a Tool model, we will return the model's name rather than
it's numeric id to make using the data with other API endpoints easier.

Bug: T274020
Change-Id: I09a9a242c6bf9a53386a2eff7762c46fc524ce6f