search: mirror toolinfo records into Elasticsearch
* Add elasticsearch-dsl, django-elasticsearch-dsl, and
  django-elasticsearch-dsl-drf libraries.
* Add django_elasticsearch_dsl integration for our
  toolhub.apps.toolinfo.models.Tool model.
* Add settings for describing how to communicate with Elasticsearch.
* Add Makefile target for creating/rebuilding Elasticsearch indices.
* Add envars for test runner containers disabling automatic replication
  of documents into Elasticsearch when model change events fire.

This initial integration dumps our Tool model into an Elasticsearch
instance in a nearly 1:1 structure. Notably there has been very little
attention paid to the Elasticsearch mapping itself at this point.
Further tuning is likely to be desired as we start to make use of the
index and discover needs for different analyzers for various fields.

Bug: T271374
Change-Id: Icc4f3fa842cf225310fb603fc4c6e7fa84904375