demo: add demo server config
Add contrib/demo-server files which can be used to run a demonstration
server from a Cloud VPS instance. The docs are a bit sketchy at this
point, but hopefully they are better than nothing for reminding the next
person how to setup the base server for hosting the demo containers.

Bug: T276357
Change-Id: I952855fd8026f2860a38e651ba32e2ba9237b7e8