toolinfo: move toolinfo.icon transforms to frontend
* Remove custom CommonsFileOutputSerializer from backend
* Add new CommonsImage component to frontend

Make working with records coming from /api/tools/ and /api/search/tools/
more similar by removing the custom output expansion for `icon` from
/api/tools/.

Replace it with a Vue component which knows how to transform a commons
File: url into direct links to rendered thumbnails for the file. This is
accomplished by calling the Commons Action API to fetch image info for
the file page. This is actually nicer than the prior use of
Special:FilePath. The API returns urls for rendered thumbnails rather
than raw files which could save a lot of network bandwidth in the case
of a large raster source image. Information is also provided to allow
srcset functionality for higher density displays.

Bug: T277235
Change-Id: I50c3a910126ccf2efe2a108141a84049e20d54ed