Soft delete for toolinfo records
Use the django-safedelete library to implement "soft" deletion for
toolinfo records. Rather than removing data from the database, this soft
delete sets a deletion date on the record. Records with a non-null
deletion date are not selected by normal operations, but may be found
using their primary key or by using special query entrypoints which
either include all records in the table or limited to only records with
a deletion date.

* update auditlog signal handlers to treat softdelete as delete
* update search signal handlers to treat softdelete as delete

Bug: T277810
Change-Id: I419c341618f6576da15f791a74c01be63efcd49a