crawler: reject updates from multiple urls in single run
The `name` field of a toolinfo record is intended to be universally
unique. This constraint is enforced by the Toolhub database. We are
also enforcing that the record for a given name is only changed by the
crawler, but currently it is possible for multiple crawled urls to
include records with the same name. This leads to "edit warring" between
the urls which is not desired.

Collect all tool names seen during a single run and log and ignore the
2nd..Nth appearances of the same name during the run. This behavior makes
the order of urls crawled significant. This may turn out to be a really
bad idea in the future, but for now it solves the edit war problem.

Bug: T278065
Change-Id: I3c08a5c8df3a32ff5a6db398fd1f3bd6918b6787