i18n: improve description of `q` search parameter
"Query string search" was meant to hint at the use of a particular
backend feature of Elasticsearch, but that hint is mostly confusing to
readers and translators. The MediaWiki search API uses "Search string"
for the closely related concept of full text search using an unspecified
query language, so let's try that. Nicely enough this is the same string
that [[User:Verdy p]] came up with when reporting the ambiguity of the
original phrase for translation.

Bug: T278170
Change-Id: I725c83814b96534f92aff9121480fa3d6a60cdfc