crawler: Capture logs from url processing
Use the new toolhub.apps.crawler.logging.CaptureCrawlLogs context
handler to collect and save logger messages emitted during processing of
each URL in a crawler run. These log events are stored in the `logs`
TextField of each RunUrl.

The UI is updated to show logs as expandable details for each crawled
URL. Display of the URL's create date has also been removed in an
attempt to reduce visual clutter in the table.

Bug: T276411
Change-Id: I7f9ca0e944eb7573bfe83bd660a290244d13e97d