toolinfo: Add LanguageData helper class
Add a new LanguageData helper class that can be used to read
a language-data.json file from
<https://github.com/wikimedia/language-data> and provide operations
similar to the PHP and JavaScript helpers provided by that project.

The language-data.json file included in this patch is take from the
1.0.2 JavaScript library's data files. It would be more ideal to
upstream this helper class and import it to our project.

Bug: T278403
Change-Id: I61bf57f3edd3e679d13a5165162e6fac851ec2aa