toolinfo: validate and clean language codes
Add validator support for validating language codes submitted as part of
a toolinfo record. These are now used to check and remove invalid codes
found during a crawl. They are also being used directly on some, but not
yet all, underlying model fields. The missing model level validation is
for the codes embedded in url_multilingual collections. These are being
cleaned and validated for crawled toolinfo.json data, but not
strongly enforced for API submissions.

Bug: T278403
Change-Id: I0116c7f197459dd26edb4a3c365d9845552c3ea9