ui: Update Home.vue back button handling
Replace the window.onpopstate handler added to Home.vue in 411821c with
a watch on the route's page query string parameter. The prior handler is
global in scope which means that it fires on all back button actions
without regard to the vue page being navigated to/from. Using a watch
from within the Home.vue component keeps the handler scoped to that
page.

Bug: T279735
Bug: T269817
Change-Id: If62977f626ab309e0612eeda168b8d515efd6bb3