ui: Replace v-chip-group with plain div
Change usage of `<v-chip-group>` to a plain div when used to display
keyword like data on toolinfo detail and card views. A v-chip-group is
intended to be a select/multi-select component, but we were only using
it to display a list of words/phrases with no need for interactivity.
There seems to be an upstream bug in v-chip-group which can trigger js
console errors like "Uncaught TypeError: this.$refs.wrapper is
undefined". Lucky for us, we don't actually need the component and can
just use a div with some padding classes added to the `<v-chip>`
elements.

Bug: T278275
Change-Id: I2b7348c6e077d3ac50ca6de16ec9317edca52b7b