ui: Refactor RevisionsDiff.vue and add header deduplication
Refactor RevisionsDiff.vue to place more logic in computed data and less
in the template itself. This also allows us to introduce deduplication
of change labels so that adding N things to an array property does not
repeat the property's name N times in the output.

The new `changes` computed data is a list of {label, index, oldValue,
newValue} changes to display. They are computed using roughly the same
logic that was being applied to the JSON Patch operations by the
template previously. The main change is that the label is optionally
null if a change applies to the same property as it's immediate
predecessor in the list.

Bug: T279740
Change-Id: I0932d77f81c4ec5e1a50443d99fc4631ab0762c2