api: Return crawled_urls count for each crawler run
Simplify ui work by adding an annotation to our Run model containing the
count of RunUrls associated with each Run. This allows us to remove
a complicated data addition that was being done in vue by fetching the
full set of RunUrls for each and every Run just so that we could compute
this count on the client side.

Bug: T280074
Change-Id: Ib3358e8f820767a5cf49811f357fc3725a2cf56e