toolinfo: validate url_multilingual collections
Introduce a validate_url_mutilingual_list validator which can be used to
validate the structure and content of a url_multilingual collection.
This patch applies that validator to the existing url_multilingual
collections in a toolinfo record. Our friendly errors configuration is
also updated to give unique error codes to custom validators which were
missing them.

Bug: T278403
Change-Id: If81e1d369a8533bb9459745e2b80d47148204316