security: Add Google FloC out-out header to responses
Add a custom middleware class that will inject a `Permissions-Policy:
interest-cohort=()` header into every response. The `interest-cohort`
feature is a control for Google Chrome's new Federated Learning of
Cohorts (FLoC) system. The `=()` value indicates to the user-agent that
the feature should be disabled for all frames regardless of origin.

See also:
* <https://github.com/wicg/floc#opting-out-of-computation>
* <https://github.com/w3c/webappsec-permissions-policy/blob/main/permissions-policy-explainer.md>

Bug: T280278
Bug: T279804
Change-Id: I33ab28ba67ea9b943dc4be8c428a17d0fe54966d