js: Set nyc code coverage to enforcement mode
Ideally we will get to >=85% code coverage for our javascript tests.
Today we are at 78% statement coverage thanks to recent efforts to add
tests for existing files. To keep us from back sliding on that progress
this patch adjusts the enforced limits downwards to meet current
coverage and enabled enforcement.

To remind us that we would like to get to 85%, the configuration is also
updated to set the "watermark" thresholds for reporting such that any
coverage <85% will be highlighted as sub-optimal coverage in reports.

Bug: T280216
Change-Id: I8b5da39e3befe4827a6eaec5fb7aa341dec7a300