authz: pass authz data to frontend and use in vue
Use the CASL library (<https://casl.js.org/v5/en/>) to provide a rules
based authorization system in our vue frontend application. The rules
are created on the backend via a transformation of the same permissions
which are used to guard our API endpoints. These rules are delivered to
the frontend as part of the payload of the /api/user/ endpoint which is
used to inform the frontend of the identity and state of the current
user.

CASL was selected for this because it works in a very similar way to the
django-rules backend library that we chose to implement object level
permissions (T278025).

On the frontend, permission checks are done using a vue plugin which
exposes the CASL `can` function. Checks can be done for rights related
to a class of business objects:

  $can( 'add', 'toolinfo/tool' )

or for a specific object:

  $can( 'delete', url )

Bug: T282186
Change-Id: Iaad5bbdf267a590db6b3fceb917c36e28ad8382d