vue: Treat timestamps as UTC when rendering
Toolhub stores and returns all timestamps in UTC time. Previously we
were rendering timestamps in the UI in the user's local timezone. This
is neat stuff, but gets into weird edge cases that are difficult to
manage when adding features that involve backend dates. The problem
there becomes figuring out how to translate from user local time into
UTC and back in all the right places. To simplify this for now (we can
always change our minds later!) we can render timestamps in the UI in
their native UTC form. With input and output in UTC we don't have to
reason about timezone conversions at all.

Change-Id: I6d1c755d7787885bc8d1e056a98058947745d75b