api: Add endpoints for suppressing a toolinfo edit
Add API endpoints which allow members of the Administrators and
Oversighters groups to toggle the visibility of a historical revision of
a toolinfo record. When a revision is hidden normal users will be able
to see that an edit was made, but unable to see the details of who made
the edit, what they changed, and the edit summary. This type of behavior
is called "revision deletion" in MediaWiki.

* PATCH /api/tools/{tool_name}/revisions/{id}/hide/
* PATCH /api/tools/{tool_name}/revisions/{id}/reveal/

Bug: T283086
Change-Id: Ie7a417776acafd37f8d1fef3e6ea578eeb115704