ui: Handle suppressed edits in toolinfo history
A suppressed edit is a historical revision of a toolinfo record which
has been hidden from normal users by an Administrator or Oversighter.
The date, user information, and edit summary for a suppressed edit are
rendered as though wrapped in `<s></s>` strikethrough elements. The
drilldown link to the revision content is only shown to users with
elevated rights.

Bug: T283086
Change-Id: I6e0955d0e738d52078b606f2ac21eee13ed6e473