api: enforce suppression when creating diffs
Add object level permissions checks when generating diffs between two
toolinfo records to ensure that suppressed content is not leaked to
unprivledged users. These checks are made against the start/end
revisions for the diff. This means that there may be leaks of content
differences if between the start and end revisions there is an
additional un-reverted but suppressed revision.

Bug: T283086
Change-Id: I15d6e1ee5a831a7ffe45405bbaa21d9c5c1641b8