db: refactor revision support to remove monkeypatch
In b2ba616 when we introduced using the django-reversion library we
copied the monkeypatch hack method used by Striker to add new metadata
to a reversion.models.Version model. While preparing to implement
a second flag for patrol status I discovered that there is a better way
to work with the upstream library. The `reversion.add_meta()` function
allows attaching an arbitrary model to each Revision model. Using this
pattern, we can support adding any arbitrary flags and metadata we are
interested in to the Revision.

* Drop monkeypatch hacks to reversion.models.Version
* Remove database migration associated with our old monkeypatch
* Add a new RevisionMetadata model to hold flags and metadata for a Revision
* Add new toolhub.apps.versioned app to hold our RevisionMetadata model
* Update existing logic to use Version.revision.meta.* for metadata
  checks and changes.

Bug: T271370
Bug: T275229
Change-Id: I9fd1db2db281b57d98be38d4effadcae6390c6bf