api: Add toolinfo revision patrolling endpoint
Add a new PATCH /api/tools/{tool_name}/revisions/{id}/patrolled/ which
can be used by Patrollers and Administrators to mark individual
revisions as reviewed.

A `patrolled` member is also added to ToolRevisionSerializer to expose
the patrolled state of each revision.

CASL consumers should check for the 'patrol' action against the
'reversion/version' model to determine if a user has the correct
permissions to mark a revision as patrolled.

Bug: T284517
Change-Id: Ib6a3bd36dac1ff26c093030115b10cc90e036654