api: Add optional "params" to auditlog entries
Some logged actions are more complex than the original (actor, action,
target) tuples that auditlog was created with. Actions such as changing
a user's group based permissions add additional "nouns" that should be
tracked as part of the historical record. MediaWiki deals with this by
having an opaque blob of data attached to each log event which
contain arbitrary action (type in MW parlance) specific content.
Consumers of the log events can then decode this information and use it
when rendering the event for human consumption.

This commit adds a "params" JSONSchema field to the Toolhub LogEvent
model. It is initially used to document:
* the user who's rights are being changed by an ADD or REMOVE event
* the revision created by a CREATE or UPDATE event referencing
  a toolinfo record
* the tool owning a revision changed by a HIDE or REVEAL event

The storage schema can be extended in the future to document additional
details that may be necessary or convenient for rendering other types of
events.

Bug: T283118
Change-Id: I842bea8161c6f5398446f4afdb0e33389c0fd721