ui: Revamp AuditLog rendering
Refactor and extend the method used to render AuditLog events to support
more complex output and the new "params" extra data. The new LogEvent
component is our first render function based component. This approach
lets us encode the business logic for placing each templated element in
native js logic rather than a sea of `v-if` checks.

New components:
* LogEvent - the <dl> containing the entire event
* EventIcon - an icon that indicates the type of event being shown
* EventTimestamp - pretty printing for the event timestamp
* UserLink - links to User pages and optionally Talk pages

This commit also introduces our first usage of the @vue/test-utils
helper library for unit testing of components. This allows us to
virtually render components and then inspect the resulting VNode objects
which is a pretty cool way to verify output for a component with a lot
of conditional branching in it's rendering.

Bug: T283118
Change-Id: I6310b8728167b58d408b0c10d06f9dfb3d9b93ed