auditlog: expose suppressed and patrolled state of revisions
Add params for the suppressed and patrolled state of the backing
revision when an auditlog's params contain a revision id. This addition
is done at output time rather than query time, so we are not currently
able to offer an API filter to only show unpatrolled revisions through
the GET /api/auditlogs/ endpoint.

Bug: T283118
Change-Id: I2cfa7f5f406d8a17676a71e543360f81602e40dc