api: Add ToolList storage and read-only API
Add the basic storage system for lists of tools. Each list has a title,
description, and icon in addition to an ordered set of tools. Lists can
be marked as "favorites" meaning that the list is a private collection
of tools only ever visible to the creating user. Alternately, a list can
be marked as "published" meaning that the author is ready to share it
with other users. Finally, a "published" list can be marked as
"featured" which will be used in the future by users with a specific
user right to highlight lists that are of high quality.

API endpoints:
* GET /api/lists/ - Paginated retrieval of list metadata. Returns count
  of tools in list along with all list metadata and summary information
  for included tools. Query parameters can be used to limit the results to
  "featured" lists. Only "published" lists will be returned unless the
  calling user is also the author of the list.
* GET /api/lists/{id}/ - Full contents of a particular list. All list
  metadata plus an array of the full Tool details.

Bug: T271484
Bug: T271488
Change-Id: Ia44d1eb5bacc2ea4a21efaad47f7c6049e061560